pyproject.toml
[tool.poetry]
name = "SignUpIn"
version = "0.1.0"
description = ""
authors = ["Veddant Parikh <parikhvedant2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
